# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soundmark API",
    "description": "Location-based social music recommendation platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recommendations": {
      "post": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Create Recommendation Endpoint",
        "description": "Create a new recommendation\n\nRequires authentication.\n\nArgs:\n    request: Recommendation creation data\n    current_user: Authenticated user\n    db: Database session\n    \nReturns:\n    Created recommendation",
        "operationId": "create_recommendation_endpoint_api_v1_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/recommendations/{recommendation_id}": {
      "get": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Get Recommendation Detail",
        "description": "Get recommendation detail\n\n**Distance restriction**: User must be within 200m of the recommendation\nto access full details.\n\nRequires authentication.\n\nArgs:\n    recommendation_id: Recommendation ID\n    lat: User's current latitude\n    lng: User's current longitude\n    current_user: Authenticated user\n    db: Database session\n    \nReturns:\n    Recommendation detail\n    \nRaises:\n    403: If user is outside 200m radius\n    404: If recommendation not found",
        "operationId": "get_recommendation_detail_api_v1_recommendations__recommendation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recommendation Id"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90.0,
              "minimum": -90.0,
              "description": "Current user latitude",
              "title": "Lat"
            },
            "description": "Current user latitude"
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180.0,
              "minimum": -180.0,
              "description": "Current user longitude",
              "title": "Lng"
            },
            "description": "Current user longitude"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{recommendation_id}/reactions": {
      "put": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Add Recommendation Reaction",
        "description": "Add or update emoji reaction to a recommendation\n\nRequires authentication.\nUser can have only one reaction per recommendation (will update if exists).\n\nArgs:\n    recommendation_id: Recommendation ID\n    request: Reaction data with emoji\n    current_user: Authenticated user\n    db: Database session\n    \nReturns:\n    Updated reactions and user's current reaction\n    \nRaises:\n    404: If recommendation not found",
        "operationId": "add_recommendation_reaction_api_v1_recommendations__recommendation_id__reactions_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recommendation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationReactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationReactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Remove Recommendation Reaction",
        "description": "Remove user's reaction from a recommendation\n\nRequires authentication.\n\nArgs:\n    recommendation_id: Recommendation ID\n    current_user: Authenticated user\n    db: Database session\n    \nReturns:\n    Updated reactions (user_reaction will be None)\n    \nRaises:\n    404: If recommendation not found",
        "operationId": "remove_recommendation_reaction_api_v1_recommendations__recommendation_id__reactions_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recommendation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationReactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{recommendation_id}/like": {
      "put": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Toggle Recommendation Like",
        "description": "Toggle like on a recommendation (deprecated - use /reactions instead)\n\nIf user has already liked, it will unlike.\nIf user hasn't liked, it will like.\n\nRequires authentication.\n\nArgs:\n    recommendation_id: Recommendation ID\n    current_user: Authenticated user\n    db: Database session\n    \nReturns:\n    Current like status and count\n    \nRaises:\n    404: If recommendation not found",
        "operationId": "toggle_recommendation_like_api_v1_recommendations__recommendation_id__like_put",
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recommendation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaceInput": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "'google' or 'manual'"
          },
          "google_place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Place Id"
          },
          "place_name": {
            "type": "string",
            "title": "Place Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "source",
          "place_name"
        ],
        "title": "PlaceInput",
        "description": "Place information input (Google or manual)"
      },
      "RecommendationCreateRequest": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "description": "Latitude"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "description": "Longitude"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "spotify_track_id": {
            "type": "string",
            "title": "Spotify Track Id",
            "description": "Spotify track ID"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Short message about the track"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Longer note about the track"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng",
          "spotify_track_id"
        ],
        "title": "RecommendationCreateRequest",
        "description": "Request model for creating a recommendation"
      },
      "RecommendationDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "distance_meters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Meters"
          },
          "track": {
            "$ref": "#/components/schemas/TrackResponse"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "reactions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reactions",
            "description": "Emoji reactions with counts"
          },
          "user_reaction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Reaction",
            "description": "Current user's reaction emoji"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "place_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lat",
          "lng",
          "track",
          "user",
          "created_at"
        ],
        "title": "RecommendationDetailResponse",
        "description": "Detailed recommendation response (used in detail view)"
      },
      "RecommendationReactionRequest": {
        "properties": {
          "emoji": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Emoji",
            "description": "Emoji reaction (unicode or name)"
          }
        },
        "type": "object",
        "required": [
          "emoji"
        ],
        "title": "RecommendationReactionRequest",
        "description": "Request model for adding/updating emoji reaction"
      },
      "RecommendationReactionResponse": {
        "properties": {
          "reactions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reactions",
            "description": "Updated emoji reactions with counts"
          },
          "user_reaction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Reaction",
            "description": "Current user's reaction emoji"
          }
        },
        "type": "object",
        "required": [
          "reactions"
        ],
        "title": "RecommendationReactionResponse",
        "description": "Response for reaction operation"
      },
      "RecommendationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "distance_meters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Meters"
          },
          "track": {
            "$ref": "#/components/schemas/TrackResponse"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "reactions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reactions",
            "description": "Emoji reactions with counts"
          },
          "user_reaction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Reaction",
            "description": "Current user's reaction emoji"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lat",
          "lng",
          "track",
          "user",
          "created_at"
        ],
        "title": "RecommendationResponse",
        "description": "Basic recommendation response (used in list views)"
      },
...
      "TrackResponse": {
        "properties": {
          "spotify_track_id": {
            "type": "string",
            "title": "Spotify Track Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "album": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "album_cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Cover Url"
          },
          "track_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Url"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "spotify_track_id",
          "title",
          "artist",
          "id",
          "created_at"
        ],
        "title": "TrackResponse",
        "description": "Track response model"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "spotify_id": {
            "type": "string",
            "title": "Spotify Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "spotify_id",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Response model for user information"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}